+++ packages/base/src/features/labels/__tests__/drawDefaultAttributes.spec.ts
