+++ packages/base/src/features/labels/drawDefaultAttributes.ts
