+++ packages/base/style/base.css
@@ -19,7 +19,7 @@
 @import url('./storySpectaArticleOverlay.css');
 @import url('./identify.css');
 @import url('./shared/propertyRows.css');
-@import url('./drawDefaultAttributesDialog.css');
+@import url('./labels.css');
 @import url('./openeo.css');
 @import url('./shared/button.css');
 @import url('./shared/toggle.css');
